Workaround (fix?) broken pathInc() call in Overview vignette
Addresses #19

Remove disconnected patches in the raster. The clump number is specific
to this raster and may not be the same for other rasters.

I'm not clear yet if this is actually necessary or is just a workaround
for a different issue. It's possible that disconnected (unreachable)
patches in the transition matrix is not possible with the matrix
decomposition, but I haven't determined yet why it used to work. It's
possible a bug was fixed in the Matrix package that was allowing the
matrix to be incorrectly decomposed, or it's possible a new bug was
introduced to the Matrix package that breaks the decomposition in
certain circumstances.